Aggiornata libreria api alla versione 0.5

+++ app/src/main/java/it/baratta/giovanni/habitat/notificator/clientdemo/service/MqttForegroundService.kt
@@ -175,7 +175,7 @@ class MqttForegroundService : Service() {
         builder.extras.putString(NotificationReciever.PAYLOAD, sb.toString())
 
         val nManager = this.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-        nManager.notify(msg.id, builder.build())
+        nManager.notify(msg.id.toInt(), builder.build())
     }
 
     companion object {
